+++ assets/javascripts/kitten/karl/tours/tour.js
@@ -1,6 +1,10 @@
 import React from 'react'
 import classNames from 'classnames'
 import { Button } from 'kitten/components/buttons/button'
+import { Marger } from 'kitten/components/layout/marger'
+import { Title } from 'kitten/components/typography/title'
+import { Paragraph } from 'kitten/components/typography/paragraph'
+import { Text } from 'kitten/components/typography/text'
 import { CloseButton } from 'kitten/components/buttons/close-button'
 import { ArrowIcon } from 'kitten/components/icons/arrow-icon'
 import { MixIllustration } from 'kitten/components/illustrations/mix-illustration'
@@ -14,30 +18,44 @@ export const KarlHeaderTour1 = props => (
     aria-labelledby="dialogtitle"
   >
 
-    <div
-      className={ classNames('k-Popover__container',
-                             'k-Tour__popover__container') }
-    >
+    <div className={ classNames('k-Popover__container',
+                                'k-Tour__popover__container') }>
 
       <div className="k-Tour__popover__illustration">
         <MixIllustration className="k-Tour__popover__svg--primary" />
       </div>
-      <div className="k-Tour__popover__content">
-        <p
-          id="dialogtitle"
-          className="k-Tour__popover__title"
-        >
-          What‘s new
-        </p>
-        <p className="k-Tour__popover__text">
-          Lorem ipsum dolor sit amet, consectetur
-          adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore
-          magna aliqua.
-        </p>
+      <Marger
+        top="4"
+        className="k-Tour__popover__content"
+      >
+        <Marger bottom="1">
+          <Title
+            tag="p"
+            modifier= "senary"
+            id="dialogtitle"
+            margin={ false }
+          >
+            What‘s new
+          </Title>
+        </Marger>
+        <Marger top="1">
+          <Paragraph
+            margin={ false }
+            modifier="quaternary"
+          >
+            Lorem ipsum dolor sit amet, consectetur
+            adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore
+            magna aliqua.
+          </Paragraph>
+        </Marger>
         <div className="k-Tour__popover__navigation">
-          <p className="k-Tour__popover__numbering">
+          <Text
+            className="k-Tour__popover__step"
+            size="tiny"
+            weight="regular"
+          >
             1 on 2
-          </p>
+          </Text>
           <div className="k-Tour__popover__buttons">
             <Button
               title="Précédent"
@@ -67,10 +85,11 @@ export const KarlHeaderTour1 = props => (
             </Button>
           </div>
         </div>
-      </div>
-      <div className="k-Popover__close">
+      </Marger>
+      <div>
         <CloseButton
-          modifier="hydrogen"
+          modifier="beryllium"
+          size="tiny"
         />
       </div>
     </div>
@@ -85,28 +104,41 @@ export const KarlHeaderTour2 = props => (
     aria-labelledby="dialogtitle"
   >
 
-    <div
-      className={ classNames('k-Popover__container',
-                             'k-Tour__popover__container') }
-    >
+    <div className={ classNames('k-Popover__container',
+                                'k-Tour__popover__container') }>
 
       <div className="k-Tour__popover__illustration">
         <LoudspeakerIllustration className="k-Tour__popover__svg--primary" />
       </div>
-      <div className="k-Tour__popover__content">
-        <p
-          id="dialogtitle"
-          className="k-Tour__popover__title"
-        >
-          What‘s new
-        </p>
-        <p className="k-Tour__popover__text">
-          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor.
-        </p>
+      <Marger
+        top="4"
+        className="k-Tour__popover__content"
+      >
+        <Marger bottom="1">
+          <Title
+            modifier="senary"
+            margin={ false }
+            id="dialogtitle"
+          >
+            What‘s new
+          </Title>
+        </Marger>
+        <Marger top="1">
+          <Paragraph
+            modifier= "quaternary"
+            margin={ false }
+          >
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor.
+          </Paragraph>
+        </Marger>
         <div className="k-Tour__popover__navigation">
-          <p className="k-Tour__popover__numbering">
+          <Text
+            className="k-Tour__popover__step"
+            size="tiny"
+            weight="regular"
+          >
             1 on 2
-          </p>
+          </Text>
           <div className="k-Tour__popover__buttons">
             <Button
               title="Précédent"
@@ -136,9 +168,12 @@ export const KarlHeaderTour2 = props => (
             </Button>
           </div>
         </div>
-      </div>
-      <div className="k-Popover__close">
-        <CloseButton modifier="hydrogen" />
+      </Marger>
+      <div>
+        <CloseButton
+          modifier="beryllium"
+          size="tiny"
+        />
       </div>
     </div>
   </div>